readyset-data: Mark normalized string types unsupported
We have a few types that we represent internally as normalized strings,
which provides a very limited type of support for them in that queries
with them may end up returning different results or results in different
orders than the upstream would. Furthermore, these types are not
currently supported in our logictests, so we are sort of blind to any
discrepancies in processing of them.

Therefore, we should consdier them as unsupported until we can properly
understand what differences there may be with the upstream types and
reason about if those differences are OK for users or not.

Change-Id: If26609f3cebc16e33f636ad3aa61917b4f0e8663