metrics: Remove unnecessary counters
Each Prometheus "summary" metric (i.e. the metric type we use for
quantiles) includes time series for the sum and count of observed
values. In many cases, we were using separate counters to count the
number or sum of observations for a histogram metric. This commit
removes these unnecessary counters to reduce the number of overall time
series being emitted by Readyset.

Change-Id: I9abdc34d6ecea6ced46212590033c9c4094b85a7