readyset-psql: Fix named_cache_queryable_after_being_cleared
This commit adds the `#[serial]` and `#[slow]` tags to the
`named_cache_queryable_after_being_cleared` test to prevent it from
running at the same time as the other tests in readyset_psql::fallback.
This fixes an issue where a table was being created by this test while
another test was running, causing the other test to fail.

Change-Id: I1676c9fcb33c1b88820dfe0706e4fa241d1c1e5c