build: Install clippy and rustfmt
For some reason, rustup was previously installing clippy and rustfmt
automatically and then suddenly stopped, breaking our build pipeline
steps that run these commands. To remediate this, this commit adds steps
in public/build/Dockerfile to manually install clippy and rustfmt, which
is what the documentation for each of these tools suggests we do.

Change-Id: I631cc6988db26fbcd3651b1bd56a0b5056edd987