readers: Fix cache miss statistics
Previously, we were always reporting 0 as the number of misses for
queries against a cache. We were passing around a `ReadReplyStats` type
that was never actually populated with statistics. This commit resolves
the issue by removing the `ReadReplyStats` type altogther and instead
reporting the number of keys missed as computed directly from the
reader.

Release-Note-Core: Fixed an issue where one of our metrics was reporting
  the incorrect number of misses for reads against caches
Change-Id: I6ed84815c222f3f4a0b8fb1f188e4966552e6633