metrics: Remove server view hit/miss metrics
This commit removes the server's and query logger's hit/miss metrics
and adds a new "hit_or_miss" label to the metric we use to measure a
query's execution time. The server's miss counter was duplicating the
the query log missed metric from the adapter, and with our previous
setup, we had no way of viewing the execution time of cache hits vs.
misses. The changes in this commit allow us to do that, while still
providing the ability to count the number of cache hits/misses by taking
advantage of the "_count" time series that comes with every Prometheus
histogram.

Change-Id: Ie3f34e9b10b6167e1032f9ab834bab60f6f68a30