logictests: Mark avg + range tests as failures
With range query autoparameterization working and enabled, many of our
generated logictests that involve ranges now result in query graphs that
include the range comparisons as parameters. For queries that include
`AVG` computations in the column list, the graphs now include an `AVG`
post-lookup aggregation. Because we don't support averages as a
post-lookup aggregation, these tests now fail. This commit marks these
tests as failures given that range autoparameterization is on by
default.

Refs: REA-4291
Change-Id: I5d4810ac9bc68289f8a708676c20194faafc7fb5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7084
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>