readyset-psql: Stream SimpleQuery responses
When we receive a query from the postgres simple query protocol, we use
tokio-postgres's SimpleQuery structures and apis to perform the query.
This can be for queries that have a small payload or for proxied queries
that have a large result set.

`fn simple_query` (what we were using before this), collects all rows in
the result into a Vec, which we were then mapping to our own SmallVec
before returning. This meant that we buffered a lot of memory before
returning.

This changes the way we proxy selects to use a streaming
protocol--tokio-postgres's SimpleQueryStream, which means we write out
rows as we stream them in rather than buffering in readyset.

Our Response structures already stream for the postgres binary protocol,
so the main changes in this patch are to add a separate, parallel way
for us to use a SimpleQueryStream, similarly to how we have PassThrough
version of RowDescription and CommandComplete that forward along the
already-serialized data from tokio-postgres.

Since the streaming was only used to provide the responses for a single
select (or execute) before, but SimpleQueryStream allows pipelining of
several responses (each will have a CommandComplete and potentially data
rows preceding it), this patch changes the logic a bit to allow for
detecting these separate result sets and injecting the appropriate
RowDescriptions. (We were already doing this logic for the simple_query
path and aggregating it into a vec.)

We had one part of our code that relied on the response being a
SimpleQuery (load_extended_types). In order to accomodate this, I moved
load_extended_types to be a part of the backend trait and added a
separate internal way of using the previous simple query protocol for
that use case.

The lack of needing to double buffer seems to also come with a
signficant performance speedup for simple query protocol SELECTs. Some
non-scientific sampling if different result sizes seems to indicate a
20-30% speedup for smaller result sizes, but for larger ones, it grows
to 40%+. This should hopefully speed up a lot of our tests as well,
which often use `simple_query`.

The psql cli tool only uses the simple query protocol, and python
psychopg2 seems to default to it as well, so this memory reduction and
speedup could be a significant win for a large set of clients.

Fixes: REA-4114
Release-Note-Core: Readyset now streams proxied results for
  postgres for queries using the simple query protocol similarly to how it
  already does for the extended query protocol. This vastly reduces the
  memory consumption for text-protocol queries as well as improves
  performance of them in many cases.
Change-Id: Ief68ecbb9252199d7e0e156b3938cb48cbe6ad7a