mysql-srv: add support to COM_CHANGE_USER
Add support for COM_CHANGE_USER command. This command is used to
change the user of the current connection.
As part of this work, we now store the client capabilities and the
initial scramble auth_plugin_data we sent to this client. This info
is used to re-authenticate the client when the COM_CHANGE_USER
command is received.

Refs: REA-4212

Release-Note-Core: Add support for MySQL COM_CHANGE_USER
  command.

Change-Id: I2790da03cefc04e21b487979738554aa3e1a2227