memory state: Simplify lookup into other indices
When we perform a lookup into a memory state and miss on the state
associated with the given columns, we attempt to perform lookups on
other states that are indexed on subsets of those columns. The logic for
these lookups was pretty complicated, handling cases individually rather
than generally. The logic also didn't handle lookups on sets of columns
that contain duplicates of the columns indexed by other states. For
example, a lookup on [0, 0] did not attempt a lookup on [0].

This commit simplifies the existing logic by generalizing it to try
every other state that is indexed on a subset of the columns for which
the lookup was initially attempted. This has the side effect of allowing
misses on keys with duplicate columns to attempt lookups on states that
are indexed on de-duplicated versions of the same column list. That last
point is important as it will be required for us to support range
queries.

Change-Id: Ibfcb8ab1726db4486183d21c392d44d89d326c03