state: Fix range lookups on persistent indexes
When performing range lookups on persistent state with ranges that have
an exclusive lower bound, we iterate over the keys in sorted order until
we find the first key that doesn't match the lower bound (since this
would be the first record in our result set). However, we were
previously comparing the *value* of the iterator to the lower bound of
the range. This was incorrect: lookups happen on *keys* and not
*values*, so we should have been comparing the *key* associated with the
iterator with the lower bound of the range.

This commit fixes the issue by ensuring that we compare to the
iterator's key when seeking for the lower bound of a range.

Release-Note-Core: Fixed an issue where range queries returned incorrect
  results in certain situations
Change-Id: If42e91237d0811920b4611f84d376cb87e931e4c