arbitrary: Lower the max year for arbitrary date
In `arbitrary::arbitrary_naive_date()`, the maximium year that we
currently generate is 262143. This is assumably because the
documentation for `NaiveDate` [0] states that it "Allows for every
proleptic Gregorian date from Jan 1, 262145 BCE to Dec 31, 262143 CE".

While one can represent a date within those year spans, one cannot
explicitly create one with those exact values. Internally,
`NavieDate` has a `MAX_YEAR` value that evaluates to 262142, one
less than the stated 262143. The code comments on that field state:

"MAX_YEAR is one year less than the type is capable of representing.
Internally we may sometimes use the headroom, notably to handle
cases where the offset of a `DateTime` constructed with
`NaiveDate::MAX` pushes it beyond the valid, representable range."

In other words, we can't pass 262143 as a year when creating
arbitrary `NaiveDate` instances. Thus, this patch sets the upper
bound to 262142.

[0] https://docs.rs/chrono/latest/chrono/struct.NaiveDate.html

Change-Id: I0d6d7955b4ccb19a0f41f284949fb3f81560f73e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7103
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>