controller: Add debug assert for reader_processing
When we ensure there is a reader as a child of a node, we first check
our readers map to see if there is already one. This commit adds a debug
assert to ensure that if there is one, it also has the same
`reader_processing` semantics and name. I think this won't matter until
we do graph re-use, but it will hopefully prevent some bugs.

Change-Id: I02d5bedfce0b9e9e3753c733cf12281854b4fd32
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7113
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>