adapter: Report correct status for Database Connection
The status reporter issued when executing SHOW READYSET STATUS was
reporting the wrong database connection status when reconnecting
to the database. The function self.upstream.is_connected() always
return an OK(bool). In case of the connection been lost, it will
report OK(false) and current code at upstream_reachable will return
false and never attempt to reconnect.

The second issue was when reconnecting. The function
self.upstream.connect() will replace the current upstream object with
a new connection object and return OK even if the connection was not
established. The status reporter was returning OK(true) without
checking if the new connection was successful.

Integration testing was booting a readyset without fallback which
caused it to not connect to MySQL when executing SHOW READYSET STATUS.

Release-Note-Core: Fixed issue on SHOW READYSET STATUS reporting
  the wrong status when reconnecting to the database.

Change-Id: I593aa4d5a161ef205ee23a3a23cab1cb84a4d6ed
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7126
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>