util: Add BoundRange
This commit adds a new `BoundRange` type to encapsulate the ranges we
use to lookup keys in B-Tree indexes. Currently, `BoundRange` just
contains two `std::ops::Bound`s, but in a future commit, that `Bound`
type will be replaced with a custom `Bound` enum without an `Unbounded`
variant. This sets the stage for us to fix issues relating to range
queries that errantly include NULL values in result sets: we should
*never* include true unbounded bounds on a range key because NULL values
should never be considered "included" in a range key.

Change-Id: I2efbd9fffeed021d5b035d65ab828493256bad20