readyset-util: Add `BoundedRange`
This commit adds a new `BoundedRange` type, which represents ranges that
are necessarily bounded. This will be used in a future commit to ensure
that we never include unbounded bounds in range keys: NULLs should never
be included in result sets returned by range keys, so we need to make
sure we include NULL as an exclusive lower bound on range keys that
"look" unbounded (e.g. x < 1).

Change-Id: I4720d6a4059c8daaa82bcb833626c91486f6124d

readyset-data: Add `IntoBoundedRange`

This commit adds a new `IntoBoundedRange` trait that makes it easy to
created `BoundedRange`s from values. The purpose of this trait is purely
to improve the ergonomics of creating `BoundedRange`s, since we cannot
use range syntax (range syntax can produce ranges that are unbounded
from above and/or below). The trait is not yet being used but will be
when we make use of `BoundedRange` in a future commit.

Change-Id: I69e6d3588768593c190940f4abc3468595240cc2

util: Add BoundRange

This commit adds a new `BoundRange` type to encapsulate the ranges we
use to look up keys in B-Tree indexes. Currently, `BoundRange` just
contains two `std::ops::Bound`s, but in a future commit, that `Bound`
type will be replaced with a custom `Bound` enum without an `Unbounded`
variant. This sets the stage for us to fix issues relating to range
queries that errantly include NULL values in result sets: we should
*never* include true unbounded bounds on a range key because NULL values
should never be considered "included" in a range key.

Change-Id: I2efbd9fffeed021d5b035d65ab828493256bad20

fix null issue

commit message

Change-Id: I7eb4dc61a5bf214c4db118ebdb05e85767201a19