readyset-util+data: Add range utilities
This commit adds a number of new utilities for working with *bounded*
ranges. A bounded range is a range whose upper and lower bounds are
never unbounded. Among the utilities added in this commit are:

- A `Bound` type very similar to `std::ops::Bound` except that it does
  not have an `Unbounded` variant
- A `BoundedRange` type alias similar to the existing `BoundPair` type
  alias except that it contains a pair of our custom `Bound` type
  instead of `std::ops::Bound`s
- A `RangeBounds` trait very similar to `std::ops::RangeBounds` except
  that the methods return our own custom `Bound` type
- An `IntoBoundedRange` trait that makes it easier to simulate unbounded
  ranges using the minimum and maximum `DfValue`s

These utilities will all be used in a future commit to ensure that we
always work with *bounded* ranges when constructing range keys.

Refs: REA-4287
Change-Id: Ibd220636bb0f3b6b9ad74eaf0db8a3192d7c96e4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7152
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>