replicators: Gracefully handle TableErrors before denying table
If we see a TableError when replicating, we drop any caches related to
that table and stop replicating that table so we can continue
replicating all the other tables. Since we don't advance the
`next_position` replication offset in the case of an error, we need to
loosen the invariant check for that lsn the first time we process a
commit after a table error.

Fixes: REA-4250
Release-Note-Core: Fixed a bug where the replicator would unnecessarily
  restart after handling an error in replication.

Change-Id: I28a977c43b73a85f9919693913aa81b8511f6c41
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7160
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>