readyset-data: Reject Numeric NaN values
A bug in rust-decimal was parsing numeric NaN values as 0s. This commit
adds a check for NaN explicitly until we can merge a fix upstream and
update our dependency on it. We already properly reject the other
special values of Infinity and -Infinity.

Refs: REA-4179
Release-Note-Core: Fixed a bug where Numeric NaN values were being
  incorrectly snapshotted as '0'.
Change-Id: I6afe98b585f60d28891b107ecd3e7f3f13237fe3