readyset-client: Refactor key construction
The code we used to construct key comparisons when building a new
`ViewQuery` was difficult to follow, relying on a very long function and
locally-defined closures. This commit factors out this code into a new
`KeyComparisonBuilder` type that is responsible for building the keys.
This should make it easier to 1) follow the code and 2) make changes to
this logic in the future.

Change-Id: Ic44514721d71691d69192e637b4780515bc9fd36