replicators: CREATE TABLE AS SELECT
Fix handling CREATE TABLE AS. This is a corner case in current pg
versions, logical message containing the DDL for the table appears
after the rows. This causes readyset to discard the write because
at the time we try to add the entries to the new table, the table has
not been created yet.
The fix ensures to resnapshot if we fail to apply one event in a table
that should be present based on readyset replication filters

Fixes: REA-4211 1165

Release-Note-Core: Fix edge case in handling CREATE TABLE AS SELECT
 for Postgres.
Change-Id: I6f16c459161e5048fefc879b787b9f62bf361af3