replicators: CREATE TABLE AS SELECT
Fix handling CREATE TABLE AS and SELECT INTO TABLE. This is a corner
case in current pg versions, logical message containing the DDL for the
table appears after the rows. This causes readyset to discard the write
because at the time we try to add the entries to the new table, the
table has not been created yet.
The fix ensures to resnapshot if we fail to apply one event in a table
that should be present based on readyset replication filters

Fixes: REA-4211
Close #1165

Release-Note-Core: Fix edge case in handling CREATE TABLE AS SELECT and
 SELECT INTO TABLE for Postgres.
Change-Id: Ia8f5b76f67ec5fe510c3c573978a377057f9aa30
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7178
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>