mysql-srv: Add CONNECT_WITH_DB capability
If the client replies to an initial handshake packet with
CapabilityFlags::CLIENT_CONNECT_WITH_DB set, RS attempts to parse the
database. This client flag is set if the client wants to connect to a
specific database after the initial handshake. In order for this
functionality to work, the server must also support the capability and
send it as part of the initial handshake packet.
We are currently not sending the CONNECT_WITH_DB capability in the
initial handshake packet and attempting to parse the database name if
the client sends the capability. This is causing the handshake to parse
the next field in the handshake packet as the database name.

Fixes: REA-4290
Close #714, close #1186

Release-Note-Core: Add the capability to connect to a specific MySQL
  database as part of the connection.
Change-Id: Ic9c511a3776f0c45bb04f0f2953f188cfa64ac65
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7179
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>