adapter: DEALLOCATE stmts while in transaction
The existing `DEALLOCATE` support missed the case of a DEALLOCATE
being requested while a transaction is open. Becuase of that, the
adapter will naively forward along the request _with the
prepared statement id the application provided_. We can't pass along
that id as we have an alternate id when we prepared the statment,
and that's the one we must send to the upstream.

This CL moves the `SqlQuery::Deallocate` statement handling slightly
higher in `Backend::query()`, just before where it checks if it's in
a transaction. Also, cleans up the existing handling  in
`Backend::query_adhoc_non_select()` as it should not execute there
any longer (but we need to handle the enum variant nonetheless).

Fixes: REA-3983
Release-Note-Core: Fixed a bug with handling
  DEALLOCATE statements within a transaction.
Change-Id: Iccfcc8ebfc8cf50840659f487e11d9099720a576
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7184
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>