mysql-srv: Adjust salt generation function
The salt generation function was generating a salt with a range between
0 and 255. MySQL requires the initial salt to be between 0 and 127
(ascii printable characters).
It also does not allow the salt to contain 0x00 (\0) or 0x24($).

This fix allows JDBC to connect to Readyset as it seems that this
driver is more strict with regards to the salt generation.

This patch is inspired by the same function MySQL uses to generate the
salt `generate_user_salt`.

Ref: REA-1135 REA-3268 REA-3339

Release-Note-Core: Adjusted the MySQL initial handshake salt generation
  function to restrict the range of random characters. This fixes a JDBC
  driver issue that randomly fails to connect to Readyset.

Change-Id: Ia629a2bf5ae7f5ae5ac2ee01544c24b623257e1b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7189
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>