readyset-psql: Mark fallback test as serial/slow
The tests in fallback.rs are serialized due to contending on the same
upstream database and are marked as slow so they don't run in CI by
default.

One of the tests was not marked as such and is likely the cause of
some flakiness we saw with the numeric_snapshot_nan test, which fails if
there are other tables than the one it expects being replicated.

Change-Id: I136d5b453788f358491feabc798830520e00bcc6