psql-srv: Expand acceptable BIND text bool values
Our current behavior for decoding bool values in PG `BIND` messages,
using the text format, has several limitations:

1) we currently only allow the value `t` to represent a `true` value.
2) anything else is considered `false`.
- This does not account for `NULL` values
- Allows any value to be `false`, even when they should be unallowed
values (`💩` currently evaluates to `false`)

The acceptable values for a bool type are listed in the PG docs[0].

We cannot solve this problem earlier, in `nom-sql` for example, as
we are in the `BIND` phase of executing a prepared statement, and
`nom-sql` typically only executes when receiving a `PARSE` or `QUERY`
message. Thus, we must handle it in the `psql-src::code::decoder`.

This CL explicitly enumerates the acceptable values for both `true`
and `false`, and will return an error on any unacceptable values.
This CL does not address the `NULL` handling issue.

[0] https://www.postgresql.org/docs/current/datatype-boolean.html

Fixes: REA-4304
Change-Id: I482a81157ac6f118dfb4ff5957aedb1285f182f9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7205
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI