util: Add fast u32 encoding
This commit adds custom functions to encode unsigned 32-bit integers as
variably-padded strings into a given `BytesMut`. Using the benchmark added
in this commit, the custom encoding performs about 18% better than Rust's
implementation of `Display` for u32.

This will be used in a future commit for improved formatting performance
of dates, times, and timestamps.

Change-Id: I0c0f3a2474421dbfb55ffb619c864753e69fb348
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7218
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>