readyset-data: Optimize TimestampTz parsing
`TimestampTz::from_str_no_bc()` parses the input string several times
over: if it can't match on finding a timestamp with time zone, it'll
reparse to find a timestamp (without time zone). If it can't find
that, it'll reparse for a date, and so on.

This CL inverts that ordering by parsing incrementally: first, find
the date. Then if more input remains, parse for a time. Then, if more
input remains, parse for a time zone (offset). We need additional
handling for the offset if the input only provided an hour offset, as
`FixedOffset::from_str()` expects hour and minute format.

Included is a benchmark test, comparing the difference in performance
for the following types:

Date:
    345.65 ns -> 110.34 ns (~70% lower)
Timestamp without offset:
    486.96 ns -> 237.06 ns (~50 lower)
Timestamp with offset:
    277.22 ns -> 239.26 ns (~14% lower)

Change-Id: I1895ed263d490ab166c346aeb033961e56e947e4