build: Use fat LTO for release binaries
In a prior commit, we update the release build profile to use "thin"
link time optimization, which improved performance substantially with no
additional build time. Compiling with "fat" LTO seems to improve
performance by an additional 1-5%, but on my machine, the builds take 3
to 4 times longer. For the majority of use cases, fat LTO is probably
unnecessary, but for release binaries, we might as well try to get the
very best performance we can.

Accordingly, this commit adds a new `release-fat-lto` build profile and
updates the buildkite steps that compile cloud binaries to use this
profile. It also updates the `release-dist` build profile to inherit
from the `release-fat-lto` profile to ensure that the binaries we
distribute also include these additional optimizations.

Change-Id: I304616308cbbfde4933555bc69fd8f0bf7da6f91
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7217
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>
Reviewed-by: Jason Brown <jason.b@readyset.io>