fallback-tests: Fix numeric_snapshot_nan assertion
The assertion in this test was assuming that there are only tables
created by the test itself, but adding some debug statements showed that
there was a 'cats' table from another test still around. Until we can
isolate the tests more, just filter out the relevant row for this
assertion to make the test not flaky.

Change-Id: I1ecd5cbcea02c3801f70e38d0d586926af2f6afa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7234
Reviewed-by: Ethan Donowitz <ethan@readyset.io>
Tested-by: Buildkite CI