adapter rewriter: Fix operand reorder on Ordering op
When the adapter rewrites a query, and discovers a `BinaryOp` with a
literal on the left and a column on the right, we swap the operands in
`readyset-sql-passes::AnalyzeLiteralsVisitor::visit_expr()`. This is
fine when the op (`BinaryOperator`) is not an ordering
comparator (like Equal).

However, if the op is ordering (like Less, LessOrEqual, ...), we need
to flip the operator to the opposite. We are not currently doing that,
and this CL flips the operator.

Fixes: REA-4193
Change-Id: I4ed455df600ab0a2b022990c2933498e05310948
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7242
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>