nom-sql: Add support for index hints
This commit adds support for index hints to the SQL parser. Index hints
are a MySQL-specific feature that allows the user to specify which index
to use or ignore for a query.
Readyset doesn't require those, but we should be able to parse and
treat any query if a variant of index hint as the same query as if
it didn't have any index hint.

Refs: REA-4320

Change-Id: I25938a3a549a28fc768a8b09f3d68a0327e1eb11
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7274
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>