replicators: Add support for Binlog Transaction Compression
Add support for MySQL Binlog Transaction Compression. This feature
allows MySQL to compress the binlog events that are sent to the
replicator. The replicator will decompress the binlog events and
process them as if they were not compressed.
The list of binlog events that can be compressed is and RS action are
as follows:
  QUERY_EVENT, // Only allows BEGIN and COMMIT.
  WRITE_ROWS_EVENT,
  UPDATE_ROWS_EVENT,
  DELETE_ROWS_EVENT

Refs: REA-4233
Closes: #747 #634

Release-Note-Core: Added support for MySQL Binlog Transaction
   Compression. This feature allows MySQL to compress the binlog
   events requiring less network bandwidth and storage space. readyset
   is now able to decompress the binlog events and process those events.

Change-Id: I367117b8cb1204e5cb86a5075ad895898656db43
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7328
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>