tests: Fix grafana test
The grafana test was still pulling readyset:lastest from ECR, but that
was out of date and failing due to a parameter change. latest-nightly is
the correct tag now.

The test was failing but reporting success, so this change also fixes
that detection as well.  However, the grafana test runs still fail at
this time, so they're being set to "soft fail" so that this change can
be merged and the tests addressed at a later time.

Change-Id: I330d81d5b8aac5dddb1fdec9c0446df1e0eb1169
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7341
Tested-by: Buildkite CI
Reviewed-by: Ron Hough <ron@readyset.io>