quickstart: Fix podman support
Previously, our quickstart compose files were not working with `podman`
because the literal `true` in yaml was being converted to a boolean
value in Python instead of a string, which leads to env vars being set
like `ENV_VAR=True`, since the Python literal for "true" is `True`. To
fix this, this commit wraps the boolean literals in our compose files in
quotes.

Change-Id: I3ca1c9bc180e8df2a6cedf3f3208a34828939b35