readyset-mysql: Disable statement cache
mysql_async has a "statement cache" feature, which limits the number of
prepared statements that exist on a given connection at one time. This
is enabled by default with a value of 32. When the number of prepared
statement exceeds 32, a previously-created prepared statement is
deallocated, which results in errors when a client attempts to execute
it.

This commit disables this statement cache to ensure that we can support
an arbitrary number of prepared statements (as we do for Postgres).

Release-Note-Core: Fixed an issue where the number of prepared
  statements for a MySQL connection was limited to 32
Change-Id: I1050e2df98a95a9d8c73c54b957ffd65e381b87b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7357
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>