readyset-psql: Use new table/enum in types::enums()
We are (yet again) still hitting a race in the `types::enum()` test
between dropping a table and enum, recreating them, and being able to
use them from within the readyset test. This patch eschews the drop
and recreate for just creating new a table and enum (for the subset of
this long test function that actually will use them).

Change-Id: Ib24914e52f25d523a88c51d15b1baf420db4caeb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7403
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>