replicators: Add support for MySQL 8.4
As part of MySQL 8.4 release, terminology about Master/Slave has been
replaced. During snapshot, we issue SHOW MASTER STATUS to gather the
current binlog position. This command has been replaced with SHOW
BINARY LOG STATUS by mysql/mysql-server@6e2c577 . Unfortunately, the
new terminology is not available on 8.0 series, so we need to check
the server version and conditionally adjust the query we issue.

Also, adjusted the checksum query to use the new terminology of
source_binlog_checksum is compatible with 8.0 and 8.4.

Ref: REA-4374
Closes: #1253

Release-Note-Core: Adjusted replicators terminology to be compatible
   with MySQL 8.4.

Change-Id: I2a57d07ef1a4a426efce3e1989d2d0e3436b6d52
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7449
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>