system-benchmarks: Fix error when no schema file
If you have an empty directory under the `benches_data/<db>`, you will
get a confusing error message like this:

```
No schema for benchmark {name}
```

It ends up being that you don't have as schema file (suffixed with
`.sql`) in on of the subdirectories. The error message fails to insert
the bench test name (from it's folder name), and thus unclear. This CL
just fixes that error message a bit.

Change-Id: I2f85bc43d5ce986894233bff855392fc1345b9d6