replicators: Drop leftover postgres tables earlier
If we are no longer replicating a table, e.g. due to the presence of
`--replication-tables-ignore`, we will register it as a non-replicated
relation on the controller. We will then later try to drop the base
table, but will instead only drop the non-replicated relation
registration, leaving the base table around. Additionally, if the table
was only partially snapshotted, we could later error when trying to
retrieve its replication offset after snapshotting has supposedly
finished (but we've haven't been replicating that table). With this
change, we drop leftover tables before registering non-replicated
relations, so that attempting to drop the base table actually does so.

Fixes: REA-3770
Change-Id: Ie98eccd888f3c250d18b72a799d0dcfb5622a872
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7472
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>