logictests: Move working where IN + aggs tests
Now that `8c9bc345a6a42f071bf1b621047f840eb9b31379` is committed, most
of the logictests that were under the `out-of-scope/ENG-629` directory
can be moved out so they execute with everything else. There are still
some failing tests, but those are almost all related to `DISTINCT`,
which isn't supported for where in and aggregates.

Change-Id: I181694eacf94a3cc04c46e5d8a16e479004ac361
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7478
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>
Reviewed-by: Michael Zink <michael.z@readyset.io>