controller: Reject unsupported MySQL storage engines
Some storage engines are unsupported only because we require consistent
snapshots when first replicating a table, and not every storage engine
has the MVCC we currently rely on for that consistency. In the future,
we could remove this exception by wrapping our snapshots with table
locks for non-transactional storage engines. This could be a bad user
experience if someone is connecting Readyset to a live cluster causing
it to, perhaps unwittingly, snapshot a large legacy MyISAM table and
possibly lock up part of their workload. We could add an option to allow
locking tables and default to rejection. Either way, supporting locking
would add some complexity to the snapshot code that does not seem
worthwhile absent explicit demand for this functionality. We already
lock tables to read their binlog position, but we would then have to
conditionally move the unlock to after the snapshot finishes. I believe
this would require making the `TableDumper` (or the `Future` that runs
it) conditionally take ownership over the connection that locked the
table, but only if we have the locking option enabled and are
snapshotting a table that requires it.

Arguably, this decision should be up to the replicator, not the
controller: as long as it can snapshot and forward rows for the table,
the server doesn't care how the table is persisted (or not), so we
should let the code in charge of replicating decide whether it can
replicate. However, the controller (and specifically `SqlIncorporator`)
is already where we parse, inspect, and make decisions about DDL; this
codepath is common to both initial snapshotting and coming across DDL
during replication; and saying "unsupported" here already does what we
want with regard to adding the table to the `TableFilter`. It does not
seem worth duplicating the parsing and discrimination in two places in
the replicator.

Fixes: REA-4328
Change-Id: I8fcc95af5057f534841c669b46e47bd774ceed58
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7498
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>